+++ backend/canisters/group/api/src/queries/mod.rs
@@ -1,5 +1,4 @@
 pub mod c2c_events_internal;
-pub mod c2c_export_group;
 pub mod c2c_name_and_members;
 pub mod c2c_summary;
 pub mod c2c_summary_updates;