+++ backend/canisters/group/impl/src/queries/mod.rs
@@ -1,5 +1,4 @@
 mod c2c_events_internal;
-mod c2c_export_group;
 mod c2c_name_and_members;
 mod deleted_message;
 mod events;