+++ frontend/app/src/components/App.svelte
@@ -88,10 +88,12 @@
         (<any>window).platformModerator = {
             addHotGroupExclusion,
             deleteFrozenGroup,
+            deleteMessage,
+            deleteChannelMessage,
             freezeGroup,
             removeHotGroupExclusion,
+            setCommunityModerationFlags,
             unfreezeGroup,
-            deleteMessage,
         };
         (<any>window).platformOperator = {
             setGroupUpgradeConcurrency,
@@ -140,6 +142,16 @@
         });
     }
 
+    function setCommunityModerationFlags(communityId: string, flags: number): void {
+        client.setCommunityModerationFlags(communityId, flags).then((success) => {
+            if (success) {
+                console.log("Community moderation flags updated", communityId);
+            } else {
+                console.log("Failed to set community moderation flags", communityId);
+            }
+        });
+    }
+
     function unfreezeGroup(chatId: string): void {
         client.unfreezeGroup({ kind: "group_chat", groupId: chatId }).then((success) => {
             if (success) {
@@ -148,6 +160,40 @@
                 console.log("Failed to unfreeze group", chatId);
             }
         });
+    }
+
+    function deleteChannelMessage(
+        communityId: string,
+        channelId: string,
+        messageId: bigint,
+        threadRootMessageIndex?: number | undefined
+    ): void {
+        client
+            .deleteMessage(
+                { kind: "channel", communityId, channelId },
+                threadRootMessageIndex,
+                messageId,
+                true
+            )
+            .then((success) => {
+                if (success) {
+                    console.log(
+                        "Message deleted",
+                        communityId,
+                        channelId,
+                        messageId,
+                        threadRootMessageIndex
+                    );
+                } else {
+                    console.log(
+                        "Failed to delete message",
+                        communityId,
+                        channelId,
+                        messageId,
+                        threadRootMessageIndex
+                    );
+                }
+            });
     }
 
     function deleteMessage(