+++ frontend/app/src/components/home/ChatsButton.svelte
@@ -1,9 +1,10 @@
 <script lang="ts">
     import { _ } from "svelte-i18n";
     import ChatListSectionButton from "./ChatListSectionButton.svelte";
+    import { emptyUnreadCounts } from "openchat-client";
 
     export let selected = false;
-    export let unread: number;
+    export let unread = emptyUnreadCounts();
 </script>
 
 <ChatListSectionButton on:click {selected} title={$_("chats")} {unread} />