+++ frontend/openchat-client/src/openchat.ts
@@ -374,6 +374,7 @@ import {
     unreadFavouriteChats,
     unreadCommunityChannels,
     globalUnreadCount,
+    getAllChats,
 } from "./stores/global";
 import { localCommunitySummaryUpdates } from "./stores/localCommunitySummaryUpdates";
 import { hasFlag, moderationFlags } from "./stores/flagStore";
@@ -3525,7 +3526,7 @@ export class OpenChat extends OpenChatAgentWorker {
     }
 
     registerProposalVote(
-        chatId: GroupChatIdentifier,
+        chatId: MultiUserChatIdentifier,
         messageIndex: number,
         adopt: boolean
     ): Promise<RegisterProposalVoteResponse> {
@@ -3948,6 +3949,15 @@ export class OpenChat extends OpenChatAgentWorker {
             });
     }
 
+    setCommunityModerationFlags(communityId: string, flags: number): Promise<boolean> {
+        return this.sendRequest({ kind: "setCommunityModerationFlags", communityId, flags })
+            .then((resp) => resp === "success")
+            .catch((err) => {
+                this._logger.error("Unable to set community moderation flags", err);
+                return false;
+            });
+    }
+
     setGroupUpgradeConcurrency(value: number): Promise<boolean> {
         return this.sendRequest({ kind: "setGroupUpgradeConcurrency", value })
             .then((resp) => resp === "success")
@@ -4073,7 +4083,7 @@ export class OpenChat extends OpenChatAgentWorker {
                     .concat(chatsResponse.state.communities.flatMap((c) => c.channels));
 
                 this.updateReadUpToStore(updatedChats);
-                const chats = this._liveState.myServerChatSummaries.values();
+                const chats = getAllChats(this._liveState.globalState).values();
 
                 this._cachePrimer?.processChatUpdates(chats, updatedChats);